Switch to Direct2D
It turned out that the way DWM composites an alpha-blended OGL window into the desktop uses some odd blend mode that I didn't expect and that can't be changed.
What it does seems to be similar to a normal blend op with premultiplied alpha, but with some twist (maybe to guarantee a tint? Maybe OS specific?). Didn't bother to try and reverse-engineer the math, because this will just get too annoying down the line. Already at the current stage it would make rendering properly anti-aliased fonts and lines much more cumbersome. In fact, I think it may not be doable at all (at least for lines where AA is out of our hands) without custom shaders, which is something I wanted to avoid for this project in the name of simplicity.

So instead this change switches things over to Direct2D, which can properly do alpha blending into the desktop with straight-alpha API inputs (even if the target formats have to use premultiplied).  It also lets us use DirectWrite which is kind of nice.
Was a bit of a hassle to get the composited D2D window actually working, but the basic setup is there and from here things should be pretty straightforward.