Standings WIP
Real-time standings progress. Pretty much works but probably not the best way to do it, will change positions based on where cars are during pacing etc. And not sure if the positions can be trusted (is the guy with most laps started and furthest along the track really always P1?).  Will look into other options.

Also added a debug overlay.