Fix hang during startup
Occasionally, iRacing would report its status as "in car" for a split second during iRon startup when that isn't actually true. This could cause the minimap logic to be fed garbage data, which in turn would result in an extremely long-running loop during offset computation.
The fix is to a) not use a loop to begin with, because it wasn't really necessary and b) add additional logic to try and work around iRacing's erroneous status reporting.

Potentially related to issue #2